employee_management/requirements.txt
Django==5.0.2
Pillow==10.2.0
psycopg2-binary==2.9.9
django-crispy-forms==2.1
crispy-bootstrap4==2023.1